Created Abstract API class for all Alpha Vantage calls and created classes for each type of call (daily, weekly, monthly, intraday) all are operational.
